removed styling temporarily, minor cleanup
